diffusion: minor file renames
