+++ python/sglang/multimodal_gen/configs/sample/teacache.py
@@ -3,7 +3,7 @@
 # SPDX-License-Identifier: Apache-2.0
 from dataclasses import dataclass, field
 
-from sglang.multimodal_gen.configs.sample.base import CacheParams
+from sglang.multimodal_gen.configs.sample.sampling_params import CacheParams
 
 
 @dataclass
